requirements.txt
flask
Werkzeug
pytesseract
opencv-python
pillow
numpy
sqlalchemy
sentence-transformers
openai
qdrant-client
ollama
faiss-cpu
PyPDF2
python-dotenv